+++ python/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Papers are now indexed by their bibkeys and can be retrieved via `Anthology.get_paper_by_bibkey()`.
+
 ## [0.5.1] — 2025-01-02
 
 This release changes the PyPI package name from acl-anthology-py to acl-anthology.